+++ pkg/ensign/config/testdata/partial.yaml
