+++ pkg/ensign/status.go
@@ -1,4 +1,4 @@
-package server
+package ensign
 
 import (
 	"context"
