+++ pkg/ensign/status_test.go
@@ -1,4 +1,4 @@
-package server_test
+package ensign_test
 
 import (
 	"context"
